Restructure mocks project
