+++ src/tests/Mocks/MockDefaultSettingsProvider.cs
@@ -2,7 +2,8 @@ using FileSurfer.Core;
 
 namespace Mocks;
 
-public sealed class MockDefaultSettingsProvider : IDefaultSettingsProvider
+public sealed class MockDefaultSettingsProvider : ServiceMock, IDefaultSettingsProvider
 {
-    public void PopulateDefaults(SettingsRecord settingsRecord) { }
+    public void PopulateDefaults(SettingsRecord settingsRecord) =>
+        RecordCall(nameof(PopulateDefaults), settingsRecord);
 }
